refactor(testruns): centralize TestRuns state logic into a custom hook
