+++ galasa-ui/src/components/test-runs/TestRunsTabs.tsx
@@ -6,34 +6,21 @@
 'use client';
 import { Tabs, Tab, TabList, TabPanels, TabPanel } from '@carbon/react';
 import styles from '@/styles/test-runs/TestRunsPage.module.css';
-import TimeframeContent, { calculateSynchronizedState } from './timeframe/TimeFrameContent';
+import TimeframeContent from './timeframe/TimeFrameContent';
 import TestRunsTable from './results/TestRunsTable';
 import SearchCriteriaContent from './search-criteria/SearchCriteriaContent';
 import TableDesignContent from './table-design/TableDesignContent';
-import { usePathname, useSearchParams, useRouter } from 'next/navigation';
 import { TestRunsData } from '@/utils/testRuns';
 import { useTranslations } from 'next-intl';
-import { useEffect, useMemo, useState } from 'react';
-import {
-  RESULTS_TABLE_COLUMNS,
-  COLUMNS_IDS,
-  TEST_RUNS_QUERY_PARAMS,
-  DAY_MS,
-  SEARCH_CRITERIA_KEYS,
-  DEFAULT_VISIBLE_COLUMNS,
-  MINUTE_MS,
-  HOUR_MS,
-} from '@/utils/constants/common';
+import { useMemo } from 'react';
+import { COLUMNS_IDS, TEST_RUNS_QUERY_PARAMS, TABS_IDS } from '@/utils/constants/common';
 import { useQuery } from '@tanstack/react-query';
-import { decodeStateFromUrlParam, encodeStateToUrlParam } from '@/utils/urlEncoder';
-import { TimeFrameValues } from '@/utils/interfaces';
-import { ColumnDefinition, runStructure } from '@/utils/interfaces';
-import { sortOrderType } from '@/utils/types/common';
 import { Run } from '@/generated/galasaapi';
-import { useDateTimeFormat } from '@/contexts/DateTimeFormatContext';
 import { useFeatureFlags } from '@/contexts/FeatureFlagContext';
 import { FEATURE_FLAGS } from '@/utils/featureFlags';
 import TestRunGraph from './graph/TestRunsGraph';
+import useTestRunsQueryParams from '@/hooks/useTestRunsQueryParams';
+import { runStructure } from '@/utils/interfaces';
 
 interface TabConfig {
   id: string;
@@ -50,118 +37,26 @@ export default function TestRunsTabs({
   resultsNamesPromise,
 }: TestRunsTabProps) {
   const translations = useTranslations('TestRunsTabs');
-  const router = useRouter();
-  const pathname = usePathname();
-  const TABS_IDS = ['timeframe', 'table-design', 'search-criteria', 'results', 'graphs'];
+
   const { isFeatureEnabled } = useFeatureFlags();
   const isGraphEnabled = isFeatureEnabled(FEATURE_FLAGS.GRAPH);
-  const rawSearchParams = useSearchParams();
-  const { getResolvedTimeZone } = useDateTimeFormat();
 
-  // Decode the search params from the URL every time the searchParams change
-  const searchParams = useMemo(() => {
-    const encodedQueryString = rawSearchParams.get('q');
-    if (encodedQueryString) {
-      const decodedQueryString = decodeStateFromUrlParam(encodedQueryString);
-      if (decodedQueryString) {
-        return new URLSearchParams(decodedQueryString);
-      }
-    }
-    return rawSearchParams;
-  }, [rawSearchParams]);
-
-  // Initialize selectedIndex based on URL parameters or default to first tab
-  const [selectedIndex, setSelectedIndex] = useState(() => {
-    const tabParam = searchParams.get('tab');
-    const initialIndex = tabParam ? TABS_IDS.indexOf(tabParam) : -1;
-    return initialIndex !== -1 ? initialIndex : TABS_IDS.indexOf('results');
-  });
-
-  // Initialize selectedVisibleColumns  based on URL parameters or default values
-  const [selectedVisibleColumns, setSelectedVisibleColumns] = useState<string[]>(
-    () =>
-      searchParams.get(TEST_RUNS_QUERY_PARAMS.VISIBLE_COLUMNS)?.split(',') ||
-      DEFAULT_VISIBLE_COLUMNS
-  );
-
-  // Initialize columnsOrder based on URL parameters or default to RESULTS_TABLE_COLUMNS
-  const [columnsOrder, setColumnsOrder] = useState<ColumnDefinition[]>(() => {
-    const orderParam = searchParams.get(TEST_RUNS_QUERY_PARAMS.COLUMNS_ORDER);
-    let correctOrder: ColumnDefinition[] = RESULTS_TABLE_COLUMNS;
-
-    // Parse the order from the URL parameter
-    if (orderParam) {
-      correctOrder = orderParam
-        .split(',')
-        .map((id) => RESULTS_TABLE_COLUMNS.find((col) => col.id === id))
-        .filter(Boolean) as ColumnDefinition[];
-    }
-
-    return correctOrder;
-  });
-
-  // Initialize timeframe values based on URL parameters or default to last 24 hours
-  const [timeframeValues, setTimeframeValues] = useState<TimeFrameValues>(() => {
-    const fromParam = searchParams.get(TEST_RUNS_QUERY_PARAMS.FROM);
-    const toParam = searchParams.get(TEST_RUNS_QUERY_PARAMS.TO);
-    const durationParam = searchParams.get(TEST_RUNS_QUERY_PARAMS.DURATION);
-
-    let toDate: Date,
-      fromDate: Date,
-      isRelativeToNow = false;
-    if (durationParam) {
-      // If duration is specified, calculate fromDate and toDate based on it
-      const [days, hours, minutes] = durationParam.split(',').map(Number);
-      toDate = new Date();
-      fromDate = new Date(
-        toDate.getTime() - (days * DAY_MS + hours * HOUR_MS + minutes * MINUTE_MS)
-      );
-      isRelativeToNow = true;
-    } else if (fromParam && toParam) {
-      // If no duration is specified, use the provided from/to dates
-      toDate = new Date(toParam);
-      fromDate = new Date(fromParam);
-    } else {
-      // If no from/to dates are specified, default to last 1 day (duration-based) from now
-      toDate = new Date();
-      fromDate = new Date(toDate.getTime() - DAY_MS);
-      isRelativeToNow = true;
-    }
-
-    const timezone = getResolvedTimeZone();
-    return { ...calculateSynchronizedState(fromDate, toDate, timezone), isRelativeToNow };
-  });
-
-  // Initialize search criteria based on URL parameters
-  const [searchCriteria, setSearchCriteria] = useState<Record<string, string>>(() => {
-    const criteria: Record<string, string> = {};
-    SEARCH_CRITERIA_KEYS.forEach((key) => {
-      if (searchParams.has(key)) {
-        criteria[key] = searchParams.get(key) || '';
-      }
-    });
-    return criteria;
-  });
-
-  // Initialize sortOrder based on URL parameters or default to an empty array
-  // URL should look like this sortOrder?result:asc,status:desc
-  const [sortOrder, setSortOrder] = useState<{ id: string; order: sortOrderType }[]>(() => {
-    const sortOrderParam = searchParams.get(TEST_RUNS_QUERY_PARAMS.SORT_ORDER);
-    let sortOrderArray: { id: string; order: sortOrderType }[] = [];
-    if (sortOrderParam) {
-      sortOrderArray = sortOrderParam.split(',').map((item) => {
-        const [id, order] = item.split(':');
-        return { id, order: order as sortOrderType };
-      });
-    }
-    return sortOrderArray;
-  });
-
-  // State to track if the component has been initialized
-  const [isInitialized, setIsInitialized] = useState(false);
-  useEffect(() => {
-    setIsInitialized(true);
-  }, []);
+  const {
+    selectedTabIndex,
+    setSelectedTabIndex,
+    timeframeValues,
+    setTimeframeValues,
+    searchCriteria,
+    setSearchCriteria,
+    sortOrder,
+    setSortOrder,
+    columnsOrder,
+    setColumnsOrder,
+    selectedVisibleColumns,
+    setSelectedVisibleColumns,
+    isInitialized,
+    searchParams,
+  } = useTestRunsQueryParams();
 
   // Define the tabs with their corresponding labels, memoized to avoid unnecessary re-renders
   const TABS_CONFIG = useMemo<TabConfig[]>(() => {
@@ -181,79 +76,6 @@ export default function TestRunsTabs({
 
     return tabs;
   }, [translations, isGraphEnabled]);
-
-  // Save and encode current state to the URL. This is the single source of truth for URL updates.
-  useEffect(() => {
-    if (!isInitialized) return;
-
-    // Build the query string from the current state
-    const params = new URLSearchParams();
-
-    // Tab
-    params.set(TEST_RUNS_QUERY_PARAMS.TAB, TABS_CONFIG[selectedIndex].id);
-
-    // Table Design
-    if (selectedVisibleColumns.length > 0) {
-      params.set(TEST_RUNS_QUERY_PARAMS.VISIBLE_COLUMNS, selectedVisibleColumns.join(','));
-    } else {
-      // If no columns are selected, we can clear the parameter
-      params.delete(TEST_RUNS_QUERY_PARAMS.VISIBLE_COLUMNS);
-    }
-    if (sortOrder.length > 0) {
-      params.set(
-        TEST_RUNS_QUERY_PARAMS.SORT_ORDER,
-        sortOrder.map((item) => `${item.id}:${item.order}`).join(',')
-      );
-    } else {
-      params.delete(TEST_RUNS_QUERY_PARAMS.SORT_ORDER);
-    }
-
-    params.set(TEST_RUNS_QUERY_PARAMS.COLUMNS_ORDER, columnsOrder.map((col) => col.id).join(','));
-
-    // Timeframe
-    if (timeframeValues.isRelativeToNow) {
-      // If relative to now, we can use the "duration" parameter
-      params.set(
-        TEST_RUNS_QUERY_PARAMS.DURATION,
-        `${timeframeValues.durationDays},${timeframeValues.durationHours},${timeframeValues.durationMinutes}`
-      );
-    } else {
-      // If not relative to now, we can use the "from" and "to" parameters
-      params.set(TEST_RUNS_QUERY_PARAMS.FROM, timeframeValues.fromDate.toISOString());
-      params.set(TEST_RUNS_QUERY_PARAMS.TO, timeframeValues.toDate.toISOString());
-    }
-
-    // Search Criteria
-    Object.entries(searchCriteria).forEach(([key, value]) => {
-      if (value) {
-        params.set(key, value);
-      } else {
-        // Remove empty criteria
-        params.delete(key);
-      }
-    });
-
-    // Encode the URL parameters to shorten the URL
-    const encodedQuery = encodeStateToUrlParam(params.toString());
-    if (encodedQuery) {
-      router.replace(`${pathname}?q=${encodedQuery}`, { scroll: false });
-    } else {
-      // If there are no params, clear the URL.
-      router.replace(pathname, { scroll: false });
-    }
-  }, [
-    selectedVisibleColumns,
-    columnsOrder,
-    sortOrder,
-    isInitialized,
-    pathname,
-    router,
-    selectedIndex,
-    searchParams,
-    timeframeValues,
-    searchCriteria,
-    TABS_CONFIG,
-  ]);
 
   /**
    * Transforms and flattens the raw API data for Carbon DataTable.
@@ -287,7 +109,7 @@ export default function TestRunsTabs({
 
   const handleTabChange = (event: { selectedIndex: number }) => {
     const currentIndex = event.selectedIndex;
-    setSelectedIndex(currentIndex);
+    setSelectedTabIndex(currentIndex);
   };
 
   // Create a canonical query key based on the search parameters so that it won't refetch unnecessarily
@@ -348,7 +170,7 @@ export default function TestRunsTabs({
       return response.json() as Promise<TestRunsData>;
     },
     // Only run the query when the results tab is selected
-    enabled: ['results', 'graphs'].includes(TABS_IDS[selectedIndex]),
+    enabled: ['results', 'graphs'].includes(TABS_IDS[selectedTabIndex]),
     // Only refetch when the canonical query key changes
     staleTime: Infinity,
   });
@@ -392,7 +214,7 @@ export default function TestRunsTabs({
   }, [runsData, sortOrder, columnsOrder]);
 
   return (
-    <Tabs className={styles.tabs} selectedIndex={selectedIndex} onChange={handleTabChange}>
+    <Tabs className={styles.tabs} selectedIndex={selectedTabIndex} onChange={handleTabChange}>
       <TabList scrollDebounceWait={200} aria-label="Test Runs Tabs">
         {TABS_CONFIG.map((tab) => (
           <Tab key={tab.label}>{tab.label}</Tab>
